requirements.txt
streamlit
supabase
google-generativeai
gtts